+++ src/main/java/org/jglr/sbm/types/MatrixType.java
@@ -2,15 +2,15 @@ package org.jglr.sbm.types;
 
 public class MatrixType extends Type {
     private final Type columnType;
-    private final int columnCount;
+    private final long columnCount;
 
-    public MatrixType(Type columnType, int columnCount) {
+    public MatrixType(Type columnType, long columnCount) {
         super("mat"+columnCount+"("+columnType+")");
         this.columnType = columnType;
         this.columnCount = columnCount;
     }
 
-    public int getColumnCount() {
+    public long getColumnCount() {
         return columnCount;
     }
